examples/experimental/werewolves/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.3
